multiplication max number of digits established

+++ Calculator2/CalculatorService/Models/MultResponse.cs
@@ -8,6 +8,6 @@ namespace CalculatorService.Models
 {
     public class MultResponse
     {
-        public int Result { get; set; }
+        public long Result { get; set; }
     }
 }
